[MOV] im_livechat: move history tracking functions out of
public_livechat.js

Task-2212347
