+++ addons/im_livechat/__manifest__.py
@@ -71,6 +71,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/tests/qunit_suite_tests/components/**/*.js',
             'im_livechat/static/src/legacy/public_livechat.js',
             'im_livechat/static/src/legacy/public_livechat_constants.js',
+            'im_livechat/static/src/legacy/public_livechat_history_tracking.js',
             'im_livechat/static/src/legacy/models/*',
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
@@ -177,6 +178,7 @@ Help your customers with this chat, and analyse their feedback.
             'mail/static/src/js/utils.js',
             'im_livechat/static/src/legacy/public_livechat.js',
             'im_livechat/static/src/legacy/public_livechat_constants.js',
+            'im_livechat/static/src/legacy/public_livechat_history_tracking.js',
             'im_livechat/static/src/legacy/models/*',
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
